Fixed errors Added stabilization by axis Z

+++ firmware/ControlBoard/MainMCU/src/smcu.h
@@ -12,7 +12,7 @@
 
 extern void smcu_init(void);
 extern void smcu_process(void);
-extern void smcu_get_sensor_data(int16_t** foot_sensors, int16_t** accel_sensor);
+extern void smcu_get_sensor_data(int16_t** foot_sensors, int32_t** accel_sensor);
 
 extern bool smcu_cli_command_process(const char* cmd, const char (*argv)[CLI_ARG_MAX_SIZE], uint32_t argc, char* response);
